+++ apps/vmq_server/src/vmq_config_cli.erl
@@ -55,7 +55,8 @@ register_config_() ->
      "topic_alias_max_broker",
      "max_last_will_delay",
      "receive_max_client",
-     "receive_max_broker"
+     "receive_max_broker",
+     "suppress_lwt_on_session_takeover"
     ],
     _ = [clique:register_config([Key], fun register_config_callback/3)
          || Key <- ConfigKeys],