+++ apps/vmq_server/test/vmq_last_will_SUITE.erl
@@ -40,6 +40,10 @@ end_per_group(_, _Config) ->
     ok.
 
 init_per_testcase(_Case, Config) ->
+    %% make sure to have sane defaults
+    vmq_server_cmd:set_config(allow_anonymous, true),
+    vmq_server_cmd:set_config(retry_interval, 10),
+    vmq_server_cmd:set_config(suppress_lwt_on_session_takeover, false),
     Config.
 
 end_per_testcase(_, Config) ->
@@ -59,11 +63,8 @@ groups() ->
          will_null_topic_test,
          will_qos0_test],
     [
-     {mqttv4, [shuffle], Tests},
-     {mqttv5, [shuffle], Tests
-     ++ [
-         will_delay_v5_test
-        ]}
+     {mqttv4, [shuffle], Tests ++ [suppress_lwt_on_session_takeover_test]},
+     {mqttv5, [shuffle], Tests ++ [will_delay_v5_test]}
     ].
 
 
@@ -218,20 +219,52 @@ assert_ge(V1, V2) when V1 >= V2 -> true.
 ts() ->
     vmq_time:timestamp(millisecond).
 
+%% TODO: port this test (and logic) to the MQTTv5 fsm as well!
+suppress_lwt_on_session_takeover_test(_Config) ->
+    enable_on_subscribe(),
+    enable_on_publish(),
+
+    LWTTopic = "suppress/will/test",
+    LWTMsg = <<"suppress-will-test-msg">>,
+
+    %% setup a subscriber to receive LWT messages
+    ConnectSub = packet:gen_connect("suppress-lwt-on-session-takeover-subscriber", [{keepalive,60}]),
+    Connack = packet:gen_connack(0),
+    Subscribe = packet:gen_subscribe(53, LWTTopic, 0),
+    Suback = packet:gen_suback(53, 0),
+    {ok, Socket} = packet:do_client_connect(ConnectSub, Connack, []),
+    ok = gen_tcp:send(Socket, Subscribe),
+    ok = packet:expect_packet(Socket, "suback", Suback),
+
+    ExpLWTPublish = packet:gen_publish(LWTTopic, 0, LWTMsg, []),
+    Connect = packet:gen_connect("suppress-lwt-on-session-takeover",
+                                 [{keepalive,60}, {will_topic, LWTTopic}, {will_msg, LWTMsg}]),
+
+    %% connect a client which won't suppress the publication of lwt messages
+    vmq_server_cmd:set_config(suppress_lwt_on_session_takeover, false),
+    {ok, _} = packet:do_client_connect(Connect, Connack, []),
+
+    %% do a session take-over and see the LWT message is published.
+    %% also suppress LWT for the new session in case of a session
+    %% take-over.
+    vmq_server_cmd:set_config(suppress_lwt_on_session_takeover, true),
+    {ok, _} = packet:do_client_connect(Connect, Connack, []),
+    ok = packet:expect_packet(Socket, "publish", ExpLWTPublish),
+
+    %% connecting again will produce no LWT message.
+    {ok, _} = packet:do_client_connect(Connect, Connack, []),
+    {error, timeout} = gen_tcp:recv(Socket, 0, 200),
+
+    disable_on_subscribe(),
+    disable_on_publish().
+
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Hooks (as explicit as possible)
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-hook_auth_on_subscribe(_,{"", <<"will-qos0-test">>}, [{[<<"will">>, <<"null">>, <<"test">>],0}]) -> ok;
-hook_auth_on_subscribe(_,{"", <<"will-qos0-test">>}, [{[<<"will">>, <<"qos0">>, <<"test">>],0}]) -> ok;
-hook_auth_on_subscribe(_,{"", <<"will-ign-sub-test">>}, [{[<<"will">>, <<"ignorenormal">>, <<"test">>], 0}]) -> ok;
-hook_auth_on_subscribe(_,{"", <<"will-delay-test-sub">>}, [{[<<"will">>, <<"delay">>, <<"test">>, <<"0">>], 0}]) -> ok.
+hook_auth_on_subscribe(_, _, _) -> ok.
 
-hook_auth_on_publish(_, _, _MsgId, [<<"ok">>], <<"should be ok">>, false) -> ok;
 hook_auth_on_publish(_, _, _MsgId, [<<"will">>,<<"acl">>,<<"test">>], <<"should be denied">>, false) -> {error, not_auth};
-hook_auth_on_publish(_, _, _MsgId, [<<"will">>,<<"null">>,<<"test">>], <<>>, false) -> ok;
-hook_auth_on_publish(_, _, _MsgId, [<<"will">>,<<"qos0">>,<<"test">>], <<"will-message">>, false) -> ok;
-hook_auth_on_publish(_, _, _MsgId, [<<"will">>,<<"ignorenormal">>,<<"test">>], <<"should be ignored">>, false) -> ok;
-hook_auth_on_publish(_, _, _MsgId, [<<"will">>,<<"delay">>,<<"test">>,<<"0">>], <<"delayed_msg">>, false) -> ok.
+hook_auth_on_publish(_, _, _, _, _, _) -> ok.
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Helper