+++ apps/vmq_server/test/vmq_publish_SUITE.erl
@@ -100,6 +100,7 @@ publish_qos1_test(Config) ->
     ok = gen_tcp:send(Socket, Publish),
     ok = mqtt5_v4compat:expect_packet(Socket, "puback", Puback, Config),
     disable_on_publish(),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket).
 
 publish_qos2_test(Config) ->
@@ -117,6 +118,7 @@ publish_qos2_test(Config) ->
     ok = gen_tcp:send(Socket, Pubrel),
     ok = mqtt5_v4compat:expect_packet(Socket, "pubcomp", Pubcomp, Config),
     disable_on_publish(),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket).
 
 
@@ -136,6 +138,8 @@ publish_b2c_disconnect_qos1_test(Config) ->
     Puback = mqtt5_v4compat:gen_puback(1, Config),
     Publish2 = mqtt5_v4compat:gen_publish("qos1/outgoing", 1,
                                   <<"outgoing-message">>, [{mid, 3266}], Config),
+    Puback2 = packet:gen_puback(3266),
+
     enable_on_publish(),
     enable_on_subscribe(),
     {ok, Socket} = mqtt5_v4compat:do_client_connect(Connect, Connack1, [], Config),
@@ -148,6 +152,8 @@ publish_b2c_disconnect_qos1_test(Config) ->
     %% send our outgoing message, when we disconnect the broker
     %% should get rid of it and ssume we're going to retry
     ok = gen_tcp:send(Socket, Publish2),
+    ok = packet:expect_packet(Socket, "puback", Puback2),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket),
     %% TODO: assert Publish2 deleted on broker
     %% reconnect
@@ -156,6 +162,7 @@ publish_b2c_disconnect_qos1_test(Config) ->
     ok = gen_tcp:send(Socket1, Puback),
     disable_on_publish(),
     disable_on_subscribe(),
+    ok = expect_alive(Socket1),
     ok = gen_tcp:close(Socket1).
 
 publish_b2c_disconnect_qos2_test(Config) ->
@@ -202,6 +209,7 @@ publish_b2c_disconnect_qos2_test(Config) ->
     ok = gen_tcp:send(Socket2, Pubcomp),
     disable_on_publish(),
     disable_on_subscribe(),
+    ok = expect_alive(Socket2),
     ok = gen_tcp:close(Socket2).
 
 publish_b2c_retry_qos1_test(Config) ->
@@ -229,6 +237,7 @@ publish_b2c_retry_qos1_test(Config) ->
     ok = gen_tcp:send(Socket, Puback),
     disable_on_publish(),
     disable_on_subscribe(),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket).
 
 publish_b2c_retry_qos2_test(Config) ->
@@ -262,6 +271,7 @@ publish_b2c_retry_qos2_test(Config) ->
     ok = gen_tcp:send(Socket, Pubcomp),
     disable_on_publish(),
     disable_on_subscribe(),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket).
 
 publish_c2b_disconnect_qos2_test(Config) ->
@@ -283,6 +293,7 @@ publish_c2b_disconnect_qos2_test(Config) ->
     ok = gen_tcp:send(Socket, Publish),
     ok = mqtt5_v4compat:expect_packet(Socket, "pubrec", Pubrec, Config),
     %% We're now going to disconnect and pretend we didn't receive the pubrec
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket),
     {ok, Socket1} = mqtt5_v4compat:do_client_connect(Connect, Connack2, [], Config),
     ok = gen_tcp:send(Socket1, PublishDup),
@@ -290,6 +301,7 @@ publish_c2b_disconnect_qos2_test(Config) ->
     ok = gen_tcp:send(Socket1, Pubrel),
     ok = mqtt5_v4compat:expect_packet(Socket1, "pubcomp", Pubcomp, Config),
     %% Again, pretend we didn't receive this pubcomp
+    ok = expect_alive(Socket1),
     ok = gen_tcp:close(Socket1),
     {ok, Socket2} = mqtt5_v4compat:do_client_connect(Connect, Connack2, [], Config),
     ok = gen_tcp:send(Socket2, Pubrel),
@@ -300,6 +312,7 @@ publish_c2b_disconnect_qos2_test(Config) ->
     end,
     ok = mqtt5_v4compat:expect_packet(Socket2, "pubcomp", Pubcomp2, Config),
     disable_on_publish(),
+    ok = expect_alive(Socket2),
     ok = gen_tcp:close(Socket2).
 
 publish_c2b_retry_qos2_test(_Config) ->
@@ -314,11 +327,10 @@ publish_c2b_retry_qos2_test(_Config) ->
     {ok, Socket} = packet:do_client_connect(Connect, Connack, []),
     ok = gen_tcp:send(Socket, Publish),
     ok = packet:expect_packet(Socket, "pubrec", Pubrec),
-    %% The broker should repeat the PUBREC
-    ok = packet:expect_packet(Socket, "pubrec", Pubrec),
     ok = gen_tcp:send(Socket, Pubrel),
     ok = packet:expect_packet(Socket, "pubcomp", Pubcomp),
     disable_on_publish(),
+    ok = expect_alive(Socket),
     ok = gen_tcp:close(Socket).
 
 pattern_matching_test(Config) ->
@@ -893,3 +905,9 @@ helper_pattern_matching(PubTopic, Config) ->
     ok = gen_tcp:send(Socket, Publish),
     ok = gen_tcp:send(Socket, mqtt5_v4compat:gen_disconnect(Config)),
     gen_tcp:close(Socket).
+
+expect_alive(Socket) ->
+    Pingreq = packet:gen_pingreq(),
+    Pingresp = packet:gen_pingresp(),
+    ok = gen_tcp:send(Socket, Pingreq),
+    ok = packet:expect_packet(Socket, "pingresp", Pingresp).